+++ gluon/gluon-config-mode-core/check_site.lua
@@ -1,0 +1,2 @@
+need_string 'config_mode.msg_welcome'
+need_string 'config_mode.msg_reboot'
