+++ gluon/gluon-config-mode-core/files/lib/gluon/config-mode/reboot/0900-msg-reboot.lua
@@ -1,0 +1,3 @@
+local site = require 'gluon.site_config'
+
+return function () luci.template.render_string(site.config_mode.msg_reboot) end
