+++ gluon/gluon-config-mode-mesh-vpn/check_site.lua
@@ -1,0 +1,1 @@
+need_string 'config_mode.msg_pubkey'
