--- gluon/gluon-config-mode/check_site.lua
@@ -1,3 +1,0 @@
-need_string 'config_mode.msg_welcome'
-need_string 'config_mode.msg_pubkey'
-need_string 'config_mode.msg_reboot'
