--- gluon/gluon-config-mode/files/usr/lib/lua/luci/view/gluon-config-mode/reboot.htm
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="<%=luci.i18n.context.lang%>" lang="<%=luci.i18n.context.lang%>">
-  <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    <title><%=hostname%> - <% if title then %><%=title%><% else %><%:Rebooting...%><% end %></title>
-    <link rel="stylesheet" type="text/css" media="screen" href="<%=media%>/cascade.css" />
-  </head>
-  <body>
-    <div id="maincontainer">
-      <div id="maincontent">
-        <h2>Geschafft! Dein Freifunkknoten ist nun fertig eingerichtet.</h2>
-        <% if pubkey then %>
-        <fieldset class="cbi-section">
-          <p>
-            <%= luci.template.render_string(site.config_mode.msg_pubkey) %>
-          </p>
-          <div class="the-key">
-            # <%= hostname %>
-            <br/>
-            <%= pubkey %>
-          </div>
-        </fieldset>
-        <% end %>
-        <fieldset class="cbi-section">
-          <%= luci.template.render_string(site.config_mode.msg_reboot) %>
-        </fieldset>
-      </div>
-    </div>
-  </body>
-</html>
